# api/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AccountDTO:
    properties:
      category:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      accounts:
        items:
          $ref: '#/definitions/dto.AccountDTO'
        type: array
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ChangePasswordDTO:
    properties:
      newPassword:
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  dto.ForgotPasswordDTO:
    properties:
      identifier:
        type: string
    required:
    - identifier
    type: object
...
  dto.LoginDTO:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  dto.LogoutDTO:
    properties:
      refreshToken:
        type: string
    type: object
...
  dto.RefreshTokenDTO:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterIndividualDTO:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - name
    - password
    - phone
    type: object
  dto.RegisterTenantDTO:
    properties:
      email:
        type: string
      ownerName:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      tenantName:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - ownerName
    - password
    - phone
    - tenantName
    type: object
  dto.ResetPasswordDTO:
    properties:
      identifier:
        type: string
      newPassword:
        maxLength: 50
        minLength: 6
        type: string
      otp:
        type: string
    required:
    - identifier
    - newPassword
    - otp
    type: object
...
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      tenantId:
        type: string
    type: object
  dto.VerifyOTPDTO:
    properties:
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: High performance Go API for AceExtension
  title: AceExtension API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change authenticated user password
      parameters:
      - description: Change Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change Password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request OTP for password reset
      parameters:
      - description: Forgot Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot Password
      tags:
      - auth
  /auth/impersonate/{tenantId}:
    post:
      consumes:
      - application/json
      description: Login as tenant owner (Super Admin only)
      parameters:
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: Impersonate Tenant
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with phone or email and password
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the session
      parameters:
      - description: Logout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Logout
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Current User
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh Token Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new tenant and its administrative user
      parameters:
      - description: Tenant Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterTenantDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new tenant
      tags:
      - auth
  /auth/register-individual:
    post:
      consumes:
      - application/json
      description: Create a new personal account and its owner
      parameters:
      - description: Individual Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterIndividualDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new individual user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with OTP
      parameters:
      - description: Reset Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset Password
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify the 6-digit OTP sent during registration
      parameters:
      - description: OTP Verification Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify registration OTP
      tags:
      - auth
...
swagger: "2.0"